+++ src/searchForm/SearchForm.unit.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { CSSTransition } from 'react-transition-group'
 import { mount, shallow } from 'enzyme'
 
 import '../../__mocks__/matchMedia'
@@ -7,13 +8,10 @@ import { MediaSize } from '../_utils/mediaSizeProvider'
 import AutoComplete from '../autoComplete/AutoComplete'
 import TextTitle from '../typography/title'
 import AutoCompleteOverlay from './autoComplete/overlay'
-import AutoCompleteSection from './autoComplete/section'
 import DatePickerOverlay from './datePicker/overlay'
-import DatePickerSection from './datePicker/section'
 import Overlay from './overlay'
 import SearchForm from './SearchForm'
 import StepperOverlay from './stepper/overlay'
-import StepperSection from './stepper/section'
 
 const today = new Date().toISOString()
 
@@ -109,27 +107,67 @@ describe('searchForm', () => {
       })
 
       it('should open the autocomplete from section', () => {
-        expect(wrapper.find(AutoCompleteSection).exists()).toBe(false)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .first()
+            .prop('in'),
+        ).toBe(false)
         wrapper.find('.kirk-searchForm-from > .kirk-search-button').simulate('click')
-        expect(wrapper.find(AutoCompleteSection).exists()).toBe(true)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .first()
+            .prop('in'),
+        ).toBe(true)
       })
 
       it('should open the autocomplete to section', () => {
-        expect(wrapper.find(AutoCompleteSection).exists()).toBe(false)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .at(1)
+            .prop('in'),
+        ).toBe(false)
         wrapper.find('.kirk-searchForm-to > .kirk-search-button').simulate('click')
-        expect(wrapper.find(AutoCompleteSection).exists()).toBe(true)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .at(1)
+            .prop('in'),
+        ).toBe(true)
       })
 
       it('should open the datepicker section', () => {
-        expect(wrapper.find(DatePickerSection).exists()).toBe(false)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .at(2)
+            .prop('in'),
+        ).toBe(false)
         wrapper.find('.kirk-searchForm-date > .kirk-search-button').simulate('click')
-        expect(wrapper.find(DatePickerSection).exists()).toBe(true)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .at(2)
+            .prop('in'),
+        ).toBe(true)
       })
 
       it('should open the stepper section', () => {
-        expect(wrapper.find(StepperSection).exists()).toBe(false)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .at(3)
+            .prop('in'),
+        ).toBe(false)
         wrapper.find('.kirk-searchForm-seats > .kirk-search-button').simulate('click')
-        expect(wrapper.find(StepperSection).exists()).toBe(true)
+        expect(
+          wrapper
+            .find(CSSTransition)
+            .at(3)
+            .prop('in'),
+        ).toBe(true)
       })
     })
   })